+++ examples/with-ethers/.env.local.example
@@ -1,7 +1,6 @@
 API_PUBLIC_KEY="<API Public Key (that starts with 02 or 03)>"
 API_PRIVATE_KEY="<API Private Key>"
 BASE_URL="https://coordinator-beta.turnkey.io"
-ORGANIZATION_ID="<Organization UUID>"
+ORGANIZATION_ID="<Organization ID>"
 PRIVATE_KEY_ID="<(Crypto) Private key ID>"
-PRIVATE_KEY_ADDRESS="<0x...>"
 INFURA_KEY="<Infura API Key>"
