+++ examples/with-ethers/package.json
@@ -3,16 +3,12 @@
   "version": "0.1.3",
   "private": true,
   "scripts": {
-    "start-eth-send": "ts-node src/scripts/eth-send.ts",
-    "start-weth-deposit": "ts-node src/scripts/weth-deposit.ts",
-    "start-weth-withdraw": "ts-node src/scripts/weth-withdraw.ts",
-    "start-weth-transfer": "ts-node src/scripts/weth-transfer.ts",
+    "start": "ts-node src/index.ts",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@turnkey/ethers": "workspace:*",
     "@turnkey/http": "workspace:*",
-    "@uniswap/sdk-core": "^3.1.1",
     "dotenv": "^16.0.3",
     "ethers": "^5.7.2"
   }
