+++ examples/with-ethers/src/weth-contract-abi.json
